Move getCoordinates to api/util/autocomplete.js and improve it
