+++ api/util/__tests__/coordinates.js
@@ -1,5 +1,5 @@
 import test from 'ava'
-import {validateCoordinatesValue, getCoordinates} from '../coordinates.js'
+import {validateCoordinatesValue} from '../coordinates.js'
 
 test('validateCoordinatesValue', t => {
   t.is(validateCoordinatesValue([2.1, 48.5]), undefined)
@@ -23,11 +23,4 @@ test('validateCoordinatesValue', t => {
     () => validateCoordinatesValue([120, 98.5]),
     {message: 'Latitude must be a float between -85 and 85'}
   )
-})
-
-test('getCoordinates', t => {
-  t.is(getCoordinates({}), undefined)
-  t.deepEqual(getCoordinates({lonlat: [2, 45]}), {lon: 2, lat: 45})
-  t.deepEqual(getCoordinates({bbox: [1, -66, 2, -66], lonlat: [2, 45]}), {lon: 2, lat: 45})
-  t.deepEqual(getCoordinates({bbox: [1, -66, 2, -66]}), {lon: 1.5, lat: -66})
 })
