+++ api/util/coordinates.js
@@ -1,6 +1,3 @@
-import center from '@turf/center'
-import bboxPolygon from '@turf/bbox-polygon'
-
 export function validateCoordinatesValue(coordinates) {
   if (!coordinates || coordinates.length !== 2) {
     throw new Error('Coordinates must be an array of two entries')
@@ -18,29 +15,5 @@ export function validateCoordinatesValue(coordinates) {
 
   if (lat < -85 || lat > 85) {
     throw new Error('Latitude must be a float between -85 and 85')
-  }
-}
-
-export function getCoordinates(params) {
-  const {bbox, lonlat} = params
-
-  if (!lonlat && !bbox) {
-    return
-  }
-
-  if (lonlat) {
-    return {
-      lon: lonlat[0],
-      lat: lonlat[1]
-    }
-  }
-
-  if (bbox) {
-    const {geometry} = center(bboxPolygon(bbox))
-
-    return {
-      lon: geometry.coordinates[0],
-      lat: geometry.coordinates[1]
-    }
   }
 }
